Сформована 0.2
