--- src/main/java/Horse.java
@@ -1,27 +1,0 @@
-public class Horse extends Herbivore {
-    public Horse() {
-        super(400, 20, 4, 60);
-    }
-
-//    @Override
-//    public void eat() {
-//        // Реалізація їжі для коня
-//    }
-
-//    @Override
-//    public void reproduce() {
-//
-//    }
-
-//    @Override
-//    public void move() {
-//
-//    }
-
-//    @Override
-//    public void die() {
-//
-//    }
-
-    // Інші перевизначені методи...
-}
