+++ src/main/java/javaRush/Carnivore.java
@@ -1,3 +1,5 @@
+package javaRush;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
