+++ src/main/java/javaRush/Horse.java
@@ -1,0 +1,8 @@
+package javaRush;
+
+public class Horse extends Herbivore {
+    public Horse() {
+        super(400, 20, 4, 60);
+    }
+
+}
