+++ src/main/java/javaRush/Main.java
@@ -1,7 +1,13 @@
 package javaRush;
 
+import org.reflections.Reflections;
+
 public class Main {
     public static void main(String[] args) {
-        // Допоможи на др
+
+
+        // Створення і запуск симуляції
+        Simulation simulation = new Simulation(100, 20, 10); // Параметри: ширина, висота острова, кількість кроків симуляції
+        simulation.start();
     }
 }
